Some documentation and refactoring
